Separated utils function from train.py. Moved example_test.py to test folder.
